+++ src/providers/twitch/TwitchMessageBuilder.cpp
@@ -43,6 +43,23 @@ const QSet<QString> zeroWidthEmotes{
     "ReinDeer", "CandyCane", "cvMask",   "cvHazmat",
 };
 
+bool isAbnormalNonce(const QString &nonce)
+{
+    // matches /[0-9a-f]{32}/
+    if (nonce.size() != 32)
+    {
+        return true;
+    }
+    for (const auto letter : nonce)
+    {
+        if (('0' > letter || letter > '9') && ('a' > letter || letter > 'f'))
+        {
+            return true;
+        }
+    }
+    return false;
+}
+
 }  // namespace
 
 namespace chatterino {
@@ -209,6 +226,28 @@ MessagePtr TwitchMessageBuilder::build()
         if (reward)
         {
             this->appendChannelPointRewardMessage(reward.get(), this);
+        }
+    }
+    if (this->tags.contains("client-nonce"))
+    {
+        auto isAbnormal =
+            isAbnormalNonce(this->tags["client-nonce"].toString());
+        if (isAbnormal && getSettings()->abnormalNonceDetection)
+        {
+            this->emplace<TimestampElement>();
+            this->emplace<TextElement>(
+                "Abnormal nonce:", MessageElementFlag::ChannelPointReward,
+                MessageColor::System);
+            this->emplace<TextElement>(this->tags["client-nonce"].toString(),
+                                       MessageElementFlag::ChannelPointReward,
+                                       MessageColor::Text);
+            this->emplace<LinebreakElement>(
+                MessageElementFlag::ChannelPointReward);
+        }
+        else if (!isAbnormal && getSettings()->normalNonceDetection)
+        {
+            this->emplace<TextElement>("*", MessageElementFlag::Text,
+                                       MessageColor(QColor(255, 163, 11)));
         }
     }